GEARs: Gear type classifications

Lists the NESFC gear types (NEGEAR) attributed to a more general "fleet" characteristic

GEARs

Format

A list of NEGEAR codes assigned to "fleets"

otter

Otter Trawls: Bottom, Bottom paired, Bottom Twin, Hadock Separator

dredge.sc

Scallop Dredges

pot

Pots and Traps: Fish, eel, conch, octopus, hagfish , shrimp, lobster, crab

longline

Bottom and Pelagic Longlines

seine

Beach, Long, mackerel, menhaden, tuna, scottish Seines

gillnet

Fixed, Anchored, Sea bass, salmon, crab, floating, run around, other Gillnets

midwater

Midwater paired and unpaired otter trawls

dredge.o

Oyster, scrapes, Clam, Crab, conch, mussel, urchin, Quahog, hydraulic dredges