Extract LOCATION information from CFDBS

Extract a list of lat, long, ten minute square, etc from the NEFSC "loc" supporting table

Usage

get_locations(channel, sqlStatement = "select * from cfdbs.loc")

Arguments

channel

an Object inherited from Oracle. This object is used to connect to communicate with the database engine. (see connect_to_database)

sqlStatement

Character string. An sql statement (optional). If no sqlStatement is provided the default sql statement "select * from cfdbs.loc" is used

Value

A list is returned:

data

containing the result of the executed sqlStatement

sql

containing the sqlStatement itself

colNames

a vector of the table's column names

Reference

Use the data dictionary for field name explanations

See also

Examples

if (FALSE) { # \dontrun{
# extracts complete locations table based on default sql statement
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
get_locations(channel)

# extracts subset of location information. Statistical area, and 10 minute square based
# on custom sql statement
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
sqlStatement <- "select area TENMSQ from cfdbs.loc;"
get_locations(channel,sqlStatement)

# extracts 10 minute square info for an area on geaorges bank (511) based on custom sql statement
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
sqlStatement <- "select area, tenmsq  from cfdbs.loc where area=511;"
get_locations(channel,sqlStatement)
} # }