Extract VESSEL information from CFDBS

Extract a list of vessell ID's, tonnage, crew size, home port, etc from the NEFSC "Mstrvess" supporting table

Usage

get_vessels(channel)

Arguments

channel

an Object inherited from ROracle::Oracle. This object is used to connect to communicate with the database engine. (see dbutils::connect_to_database)

Value

A list is returned:

data

containing the result of the executed sqlStatement

sql

containing the sqlStatement itself

colNames

a vector of the table's column names

If no sqlStatement is provided the default sql statement "select * from NEFSC_GARFO.cfdbs_mstrvess" is used

Reference

Use the data dictionary for field name explanations

See also

Examples

if (FALSE) { # \dontrun{
# extracts complete vessel table based on custom sql statement
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
get_vessels(channel)

} # }