Plots comlands data

Comland data landings or the value of landings are plotted for each EPU over time by either GEAR type, Quarter Year (QY), vessel SIZE, or location of landings (US/nonUS)

Usage

plot_comland(data, by, type = "landings", range = NA, free.y = T)

Arguments

data

Data frame. Output from comland.r

by

Character string. The variable within data that you want to plot by (GEAR, QY, SIZE, US)

type

Character string. Plot either 'landings' (default) or 'value'.

range

Numeric vector. Range of years to plot. Default value plots all years.

free.y

Boolean. Allow the y-axis to rescale per "By" group (Default = T)