soeAreas: Area Species,EPU designations

Option to pass to get_comland_data if the user wishes define EPUs exclusively by Stat areas and not to proportion the landings in these Statistical areas to EPUs

Usage

soeAreas

Format

A Data frame

AREA

Statistical Area

NESPP3

Gulf of Maine EPU

MeanProp

Proportion of Statistical area to attribute to EPU

EPU

Ecological Production Unit