Extracts commercial data from Database

Connects to cfdbs and pulls fields from WOLANDS, WODETS, CFDETS

get_comland_data(channel, landed, endyear, out.dir)

Arguments

channel
an Object inherited from DBIConnection-class. This object is used to connect to communicate with the database engine. (see connect_to_database)

landed
Character String. Use landed weight ("y" - meatwt) for scallops and clams instead of live weight ("n" - livewt).

endyear
Numeric Scalar. Final year of query.

out.dir
path to directory where final output will be saved

Value

Data frame (data.table) (n x 10) Each row of the data.table represents a species record for a given tow/trip

YEAR

Year of trip/tow

MONTH

Month of trip/tow

NEGEAR

Fishing gear used on trip/tow

TONCL1

Tonnage class of the fishing vessel

NESPP3

Species code (3 charachters)

NESPP4

Species code and market code (4 characters)

AREA

Statistical area in which species was reportly caught

UTILCD

Utilization code

SPPLIVLB

live weight (landed = "n") or landed weight (landed="y") in lbs

SPPVALUE

The value of landed catch to the nearest dollar (U.S.), paid to fisherman by dealer, for a given species.

File Creation

A file containing the data.table above will also be saved to the users machine in the directory provided